+++ src/components/controls/buyForm/buy-form.module.less
@@ -38,30 +38,47 @@
           @media screen and (min-width: @breakPointThird) {
             margin-top: 1rem;
           }
-          .quantity {
+          .quantity-wrapper {
             display: flex;
-            align-items: center;
+            height: 100%;
             margin-right: 1rem;
-            height: 100%;
-
-            .quantity-field {
-              display: block;
+            .quantity {
+              display: flex;
+              align-items: center;
               height: 100%;
-              width: 50px;
-              padding: 0;
-              text-align: center;
-              margin-right: 1rem;
-              align-self: normal;
-            }
-          }
 
-          .quantity-button {
-            height: 100%;
-            background: @linkColor;
-            // border: 1px solid @linkColor;
-            &:hover {
-              background: lighten(@linkColor, 8%);
-              border-color: lighten(@linkColor, 8%);
+              .quantity-field {
+                font-size: 1rem;
+                font-family: @defaultFont;
+                font-weight: @fontWeightMedium;
+                display: block;
+                appearance: none;
+                border: 1px solid @borderColor;
+                border-left: 0;
+                border-right: 0;
+                height: 100%;
+                width: 50px;
+                padding: 0;
+                text-align: center;
+                align-self: normal;
+                outline: none;
+              }
+            }
+
+            .quantity-button {
+              font-size: 1.5rem;
+              padding: 0;
+              width: 50px;
+              font-family: @defaultFont;
+              font-weight: @fontWeightMedium;
+              height: 100%;
+              color: #000;
+              background: #fff;
+              border: 1px solid @borderColor;
+              &:hover {
+                background: @borderColor;
+                border-color: @borderColor;
+              }
             }
           }
