+++ src/pages/404.js
@@ -1,6 +1,8 @@
 import React from "react"
 import { Helmet } from "react-helmet"
 import { siteData } from "../data/site/sitedata"
+import Grid from "../components/gridLayout/grid/Grid"
+import Widget from "../components/gridLayout/widget/Widget"
 import Structure from "../components/structure/Structure"
 import { Link } from "gatsby"
 
@@ -12,12 +14,16 @@ export default function NotFound() {
         <meta property="og:type" content="404" />
       </Helmet>
       <Structure>
-        <h1>404 Page not found</h1>
-        <h2>Hoppsan! Sidan du försökte gå till hittades inte</h2>
-        <p>
-          Gå till <Link to="/products">produkter</Link> eller till{" "}
-          <Link to="/">startsidan</Link>.
-        </p>
+        <Grid>
+          <Widget>
+            <h1>404 Page not found</h1>
+            <h2>Hoppsan! Sidan du försökte gå till hittades inte</h2>
+            <p>
+              Gå till <Link to="/products">produkter</Link> eller till{" "}
+              <Link to="/">startsidan</Link>.
+            </p>
+          </Widget>
+        </Grid>
       </Structure>
     </>
   )
